// src/app/interfaces/categoria-tarjeta.enum.ts
export enum CategoriaTarjeta {
  BAZAR = 'bazar',
  BEBIDA = 'bebida',
  CARNICERIA = 'carniceria',
  DIETETICA = 'dietetica',
  ELECTRONICA = 'electronica',
  FARMACIA = 'farmacia',
  FERRETERIA = 'ferreteria',
  FIAMBRERIA = 'fiambreria',
  HELADERIA = 'heladeria',
  JUGUETERIA = 'jugueteria',
  KIOSCO = 'kiosco',
  LIBRERIA = 'libreria',
  LIMPIEZA = 'limpieza',
  MUEBLERIA = 'muebleria',
  OTROS = 'otros',
  PANADERIA = 'panaderia',
  PESCADERIA = 'pescaderia',
  ROPA = 'ropa',
  SUPERMERCADO = 'supermercado',
  TIENDA = 'tienda',
  VERDULERIA = 'verduleria',
  VETERINARIA = 'veterinaria'
}